OPEN - issue BATCH-207: exit_message too short for holding long exceptions http://opensource.atlassian.com/projects/spring/browse/BATCH-207
Added onFlushDirty to the entity interceptor in hibernate (and changed its name so not specific to entity name - BatchHibernateInterceptor).  Also truncating in Sql*Dao.  Only exit description is given special treatment for now, and the length is fixed at 250 - hard-coded and tightly coupled with the hibernate mapping and DDL).  Should be good enough for most scenarios.