IN PROGRESS - issue BATCH-241: Support for JobLauncher.stop() on job that hasn't started yet (delayed execution) http://opensource.atlassian.com/projects/spring/browse/BATCH-241
Added some more internal bookkeeping in the JobLauncher so that it can recognise a job that it was asked to start, but hasn't been executed by the task executor yet.  A side effect is that the JobExecution for such a job is never updated, so I'll keep this issue open until that is also fixed.