Some updates to support running multiple jobs simultaneously: First of all, longs cannot be incremented (++) safely. Second, no reason to use external synchronization when the ConcurrentMap API gives you the functionality we are looking for. Finally, since we often are adding to the end (higher keys), a ConcurrentSkipListMap should be more performant than any other kind of concurrent map.
Forgot the unit test to demonstrate the id generation concurrency bug

Additional updates to support concurrency in the in-memory execution.

Also add the final keyword to the internal variable in order to allow some runtime optimizations and initialization guarnatees. Shouldn't be a big deal as things sit, but it's a good practice in general.