Update dependency versions
In addition to updating the versions of numerous dependencies, this
commit also remove the dependency upon CGLib as it is no longer used.

Two test classes have been updated. One to cope with a new method in
Commons IO FileUtils which made a method call ambiguous, and one to
cope with API changes in the latest version of Quartz.