Update configuration of Spring IO plugin
This commit updates the configuration of the IO plugin to use an
external property, platformVersion, and to only apply the plugin when
this property is set. This removes the declaration of a snapshot
dependency from the normal build configuration and also breaks a
circular dependency between the projects that are in the Platform and
the Platform itself.

The commit also applies the configuration that was applied to the
standard test task, other than code coverage, to the new test tasks
that are created by the Spring IO plugin.

Two redundant Hibernate mapping files have been deleted. With more
recent versions of Hibernate the presence of these files was causing
test failures due to duplicate mappings.