Update the tests to use the latest version of JRuby 1.x
Previously, the JRuby engine that was being used pulled in an old
version of JRuby. The engine is incompatible with more recent JRuby 1.x
releases which causes problems for the Spring IO Platform which uses
JRuby 1.7.x.

This commit updates the tests to use the latest version of JRuby 1.x
(1.7.22) which contains its own JSR 223 integration, i.e. there’s no
need for a separate script engine dependency.