Docs support toggling Java Config and XML
This is a demo of the functionality and it located in the domain.adoc file.

added code to handl PDF ifdef

JavaScript added to support slider control

If view in local file system use safari.   (Cookies not supported on local filesysem on chrome)
Remaining bug...  if user selects xml and refreshes the examples still show XML but slider goes back to JAVA.

Could use a new tool suite.  But this is based on the Spring.io toggle.

Implemented java vs xml toggle in reference documentation

This commit ports the javascript widget used in Sagan for toggling
between Maven and Gradle on each project page and uses it to toggle
between java based config and XML based config.  The preference the user
selects is cookied so that it should stick between pages in the multiple
page layout as well as for future visits to the documentation.

The only page that has implemented the toggle in this commit is the
domain.adoc page.  The spring-batch-intro.adoc has the toggle reference
in it, but it's only there to verify multiple includes are working.

Polishing