Docs support toggling Java Config and XML
This is a demo of the functionality and it located in the domain.adoc file.

added code to handl PDF ifdef

JavaScript added to support slider control

If view in local file system use safari.   (Cookies not supported on local filesysem on chrome)
Remaining bug...  if user selects xml and refreshes the examples still show XML but slider goes back to JAVA.

Could use a new tool suite.  But this is based on the Spring.io toggle.