Implemented java vs xml toggle in reference documentation
This commit ports the javascript widget used in Sagan for toggling
between Maven and Gradle on each project page and uses it to toggle
between java based config and XML based config.  The preference the user
selects is cookied so that it should stick between pages in the multiple
page layout as well as for future visits to the documentation.

The only page that has implemented the toggle in this commit is the
domain.adoc page.  The spring-batch-intro.adoc has the toggle reference
in it, but it's only there to verify multiple includes are working.