Various changes to support a better document toggle.
I got a simpler document toggle to work with the dynamic ToC, and I added cookies to remember the doc toggle setting for our readers. I started from scratch, with an eye specifically toward simplifying the HTML and JavaScript and doing more with CSS. I have tested it on Chrome, Firefox, Safari, and Opera. I'll try it on IE/Edge after this commit (when I can hit it from my Windows laptop).