BATCH-1767: fix optimistic locking exception in multi-threaded step
Currently, when the commit of the chunk fails in a multi-threaded step,
the step execution is rolled back to a previous, eventually obsolete
version. This previous version might be obsolete because it could be
modified by another thread. In this case, a OptimisticLockingFailureException
is thrown when trying to persist the step execution leaving the step in
an UNKNOWN state while it should be FAILED.

This commit fixes the issue by refreshing the step execution to the
latest correctly persisted version before applying the step contribution
so that the contribution is applied on a fresh correct state.

Resolves BATCH-1767