BATCH-2491: use LinkedHashSet instead of HashSet for command line options
Before this commit, options passed on the command line were collected
in a HashSet. This does not keep options order as they are passed in
the command line.

This commit changes the "opts" variable type to LinkedHashSet.
Note there is no test for this change as "opts" is a local variable to
the main method.

Resolves BATCH-2491