BATCH-2465: add note about the importance to have a unique key constraint on the sortKey
This PR is for BATCH-2465.

It updates the reference documentation and the Javadoc to mention
explicitly the importance to have a unique key constraint on the sortKey
when using paging item readers.

I deliberately used "It is important" and not "It is required" (or
mandatory) because in the end this is not required but in that case
there is no guarantee to not lose data between executions.

Please let me know if the wording should be updated.