BATCH-2294: Don't replace my PlatformTransactionManager with a new one
For some reason DefaultBatchConfigurer decides that you can't possibly
already have a PlatformTransactionManager and makes a new one for you.
Because of the way Spring's bean factory works, this new one is
sometimes created after the one you _actually_ wanted, and it _replaces_
the real one in your application with a lame one that is useless.

This patch changes the implementation of DefaultBatchConfigurer so it
tries to autowire an existing bean and only makes a new one if one
doesn't exist already.