[BATCH-2760] Failed JobExecution due to unavailable TaskExecutor leaves End Time unpopulated
On instances when the taskExecutor is not picking up the submitted Job Execution, with this fix we are explicitly setting the End Time of the JobExecution in order not to give false positives via the JobExplorer#findRunningJobExecutions(). This last method is judging on running Job Executions based on End Time being null.