Fixing problems in the build
Mahmoud Ben Hassine discovered that we had JavaScript problems. It turns out that we had two problems: The JavaScript files weren't loading, and the local copiy of docinfo-footer.html was masking the same-named file from spring-doc resources. I fixed all that (and made the build file a little more sane in the process).

Thanks, Mahmoud.