Fix JobOperatorFunctionalTests#testMultipleSimultaneousInstances
The `testMultipleSimultaneousInstances` test uses a SimpleAsyncTaskExecutor.
This means when a job is submitted, a new thread will be created to run
the job. However, there could be a small time interval between
JobOperator.startNextInstance(job) and
JobOperator.findRunningExecutions(job) where the job execution is created
but not started yet. When this happens, the test fails as
`findRunningExecutions` does not return the just created
(but not started yet) execution.

This commit adds a `Thread.sleep` between these two invocations in order
to give a chance to the background thread (to be created and) to execute
the job.

(cherry picked from commit 90a89c1b02edecb596005efb8ad23e144af490bb)