Fix incorrect description of AggregateItemReader in "Appendix A"
This commit fixes incorrect references to `AggregateItemReader#__$$BEGIN_RECORD$$__`
and `AggregateItemReader#__$$END_RECORD$$__` by updating the description of the
reader with the one in its Javadoc.

It also adds a note that the AggregateItemReader is not part of the standard library
of readers provided by Spring Batch but only given as a sample.

Issue #1793