Upgrade micrometer version to 1.7
Version 1.5.17 being the last OSS
release in the 1.5.x line, this commit
upgrades micrometer's version to 1.7.x
which should be backward compatible.

Since this upgrade is done in a patch
release, the version is updated to the
least supported version in order to
minimize the gap between minor versions.

While bumping minor versions in patch
releases is not typical, this is a
necessary upgrade to run a supported
version of micrometer.